Changing prefix2 to be C-b instead of C-a
From the comment, it appeared as if the intention of prefix2 was
to keep default tmux functionality for users who did not want to
use ```.  However, the default pref is C-b as determined from
the [tmux man page](https://man7.org/linux/man-pages/man1/tmux.1.html)
   
> tmux may be controlled from an attached client by using a key
     combination of a prefix key, ‘C-b’ (Ctrl-b) by default, followed by
     a command key
   
Therefore, C-a was changed to C-b and the comment updated to reflect
 this change.
   
Additionally, whitespace at the end of the files was removed.